rooms with scores of zero are now filtered out; renamed ernMatch
